+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning].
 - More reader-friendly README (#144)
 - Add author names to licenses (#145)
 
+### Changed
+
+- Signature of generate now accepts source path, which defaults to the URL (#174)
+
 ## [0.3.1] - 2024-05-23
 
 ### Added